+++ src/Tests/Testing.Common/Testing/CompilerDiagnosticFixTestData.cs
@@ -20,6 +20,7 @@ public sealed class CompilerDiagnosticFixTestData
     /// <param name="source"></param>
     /// <param name="additionalFiles"></param>
     /// <param name="equivalenceKey"></param>
+    [Obsolete("This constructor is obsolete and will be removed in future versions.")]
     public CompilerDiagnosticFixTestData(
         string diagnosticId,
         string source,
@@ -32,6 +33,24 @@ public sealed class CompilerDiagnosticFixTestData
         EquivalenceKey = equivalenceKey;
     }
 
+#pragma warning disable CS0618 // Type or member is obsolete
+    /// <summary>
+    /// Initializes a new instance of <see cref="CompilerDiagnosticFixTestData"/>
+    /// </summary>
+    /// <param name="source"></param>
+    /// <param name="additionalFiles"></param>
+    /// <param name="equivalenceKey"></param>
+    public CompilerDiagnosticFixTestData(
+        string source,
+        IEnumerable<AdditionalFile>? additionalFiles = null,
+        string? equivalenceKey = null)
+    {
+        Source = source ?? throw new ArgumentNullException(nameof(source));
+        AdditionalFiles = additionalFiles?.ToImmutableArray() ?? ImmutableArray<AdditionalFile>.Empty;
+        EquivalenceKey = equivalenceKey;
+        DiagnosticId = null!;
+    }
+
     internal CompilerDiagnosticFixTestData(CompilerDiagnosticFixTestData other)
         : this(
             diagnosticId: other.DiagnosticId,
@@ -40,10 +59,12 @@ public sealed class CompilerDiagnosticFixTestData
             equivalenceKey: other.EquivalenceKey)
     {
     }
+#pragma warning restore CS0618 // Type or member is obsolete
 
     /// <summary>
     /// Gets compiler diagnostic ID to be fixed.
     /// </summary>
+    [Obsolete("This property is obsolete and will be removed in future versions.")]
     public string DiagnosticId { get; }
 
     /// <summary>
@@ -62,7 +83,7 @@ public sealed class CompilerDiagnosticFixTestData
     public string? EquivalenceKey { get; }
 
     [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-    private string DebuggerDisplay => $"{DiagnosticId}  {Source}";
+    private string DebuggerDisplay => $"{Source}";
 
     /// <summary>
     /// Creates and return new instance of <see cref="CompilerDiagnosticFixTestData"/> updated with specified values.