+++ src/Tests/Testing.Common/Testing/RefactoringVerifier.cs
@@ -26,6 +26,66 @@ public abstract class RefactoringVerifier<TRefactoringProvider> : CodeVerifier
     /// <summary>
     /// Verifies that refactoring will be applied correctly using specified <typeparamref name="TRefactoringProvider"/>.
     /// </summary>
+    /// <param name="source">Source code where text selection is marked with <c>[|</c> and <c>|]</c> tokens.</param>
+    /// <param name="expectedSource"></param>
+    /// <param name="additionalFiles"></param>
+    /// <param name="equivalenceKey"></param>
+    /// <param name="options"></param>
+    /// <param name="cancellationToken"></param>
+    public async Task VerifyRefactoringAsync(
+        string source,
+        string expectedSource,
+        IEnumerable<string>? additionalFiles = null,
+        string? equivalenceKey = null,
+        TestOptions? options = null,
+        CancellationToken cancellationToken = default)
+    {
+        var code = TestCode.Parse(source);
+
+        var expected = ExpectedTestState.Parse(expectedSource);
+
+        var data = new RefactoringTestData(
+            code.Value,
+            code.Spans.OrderByDescending(f => f.Start).ToImmutableArray(),
+            AdditionalFile.CreateRange(additionalFiles),
+            equivalenceKey: equivalenceKey);
+
+        await VerifyRefactoringAsync(
+            data,
+            expected,
+            options,
+            cancellationToken: cancellationToken);
+    }
+
+    internal async Task VerifyRefactoringAsync(
+        string source,
+        string sourceData,
+        string expectedData,
+        IEnumerable<string>? additionalFiles = null,
+        string? equivalenceKey = null,
+        TestOptions? options = null,
+        CancellationToken cancellationToken = default)
+    {
+        var code = TestCode.Parse(source, sourceData, expectedData);
+
+        var expected = ExpectedTestState.Parse(code.ExpectedValue!);
+
+        var data = new RefactoringTestData(
+            code.Value,
+            code.Spans.OrderByDescending(f => f.Start).ToImmutableArray(),
+            AdditionalFile.CreateRange(additionalFiles),
+            equivalenceKey: equivalenceKey);
+
+        await VerifyRefactoringAsync(
+            data,
+            expected,
+            options,
+            cancellationToken: cancellationToken);
+    }
+
+    /// <summary>
+    /// Verifies that refactoring will be applied correctly using specified <typeparamref name="TRefactoringProvider"/>.
+    /// </summary>
     /// <param name="data"></param>
     /// <param name="expected"></param>
     /// <param name="options"></param>
@@ -105,6 +165,32 @@ public abstract class RefactoringVerifier<TRefactoringProvider> : CodeVerifier
                     await VerifyAdditionalDocumentsAsync(document.Project, expectedDocuments, cancellationToken);
             }
         }
+    }
+
+    /// <summary>
+    /// Verifies that refactoring will not be applied using specified <typeparamref name="TRefactoringProvider"/>.
+    /// </summary>
+    /// <param name="source">Source code where text selection is marked with <c>[|</c> and <c>|]</c> tokens.</param>
+    /// <param name="equivalenceKey"></param>
+    /// <param name="options"></param>
+    /// <param name="cancellationToken"></param>
+    public async Task VerifyNoRefactoringAsync(
+        string source,
+        string? equivalenceKey = null,
+        TestOptions? options = null,
+        CancellationToken cancellationToken = default)
+    {
+        var code = TestCode.Parse(source);
+
+        var data = new RefactoringTestData(
+            code.Value,
+            code.Spans,
+            equivalenceKey: equivalenceKey);
+
+        await VerifyNoRefactoringAsync(
+            data,
+            options,
+            cancellationToken: cancellationToken);
     }
 
     /// <summary>